+++ main.go
@@ -12,6 +12,7 @@ import (
 	_ "github.com/MustWin/cmeter/containers/embedded"
 	"github.com/MustWin/cmeter/context"
 	_ "github.com/MustWin/cmeter/reporting/cmeterapi"
+	_ "github.com/MustWin/cmeter/reporting/http"
 	_ "github.com/MustWin/cmeter/reporting/mock"
 )