+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `http` reporting driver for generic http reporting
 ### Removed
 - the `api` command.
 - `mockapi` configuration section.